% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_model_check.R
\name{interactive_model_check}
\alias{interactive_model_check}
\title{check model existance during interactive use to protect against improperly
set paths}
\usage{
interactive_model_check(.model)
}
\arguments{
\item{.model}{modelfile name in directory of file where overseer is set}
}
\description{
check model existance during interactive use to protect against improperly
set paths
}
\examples{
\dontrun{
# given a model my_model.cpp
if (!interactive_model_check("my_model.cpp")) {
  stop("make sure the directory is set to the models directory before running
  interactively to make sure the relative paths will be the same as when sourcing")
}
}
}
