% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overseer.R
\docType{data}
\name{Overseer}
\alias{Overseer}
\title{Model repository manager}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Overseer
}
\description{
Model repository manager
}
\examples{
\dontrun{
model_cache <- Overseer$new(".modcache") # relative path to where should cache files
model_cache$add_model("test")
}
}
\keyword{datasets}

