% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_models.R
\name{source_models}
\alias{source_models}
\title{source models from an overseer R script}
\usage{
source_models(.path)
}
\arguments{
\item{.path}{path to R script containing overseer setup}
}
\description{
source models from an overseer R script
}
