% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overseer.R
\name{Overseer}
\alias{Overseer}
\title{Model repository manager}
\description{
Model repository manager
}
\examples{
\dontrun{
model_cache <- Overseer$new(".modcache") # folder name for the cache files
model_cache$add_model("test")
}
}

